Fixed issue #80
The simple query build was only showing the current page's worth of data
when in edit mode (display mode was ok). I've change this to load the
full data set when a Chart has been added, and update the paging to use
that data set so multiple round trips are not required.